+++ app/src/main/java/com/worldsnas/starplayer/App.kt
@@ -1,12 +1,13 @@
 package com.worldsnas.starplayer
 
 import android.app.Application
-import com.worldsnas.starplayer.di.AppComponent
-import com.worldsnas.starplayer.di.DaggerAppComponent
+import com.worldsnas.starplayer.di.components.AppComponent
+import com.worldsnas.starplayer.di.components.DaggerAppComponent
+
 
 class App  :Application(){
 
-    lateinit var appComponent :AppComponent
+    lateinit var appComponent : AppComponent
 
     override fun onCreate() {
         super.onCreate()
