+++ app/src/main/java/com/worldsnas/starplayer/di/components/MusicListComponent.kt
@@ -1,5 +1,8 @@
-package com.worldsnas.starplayer.di
+package com.worldsnas.starplayer.di.components
 
+import com.worldsnas.starplayer.di.FragmentScope
+import com.worldsnas.starplayer.di.modules.MusicProviderModule
+import com.worldsnas.starplayer.di.modules.ViewModelModule
 import com.worldsnas.starplayer.view.musics_list.MusicsListFragment
 import dagger.Component
